Security: Cast course_id to int in model.ajax to prevent SQL injection
see advisory GHSA-pjh4-mvwg-6jrm

+++ public/main/inc/ajax/model.ajax.php
@@ -743,7 +743,7 @@ switch ($action) {
             exit;
         }
 
-        $courseId = $_REQUEST['course_id'] ?? 0;
+        $courseId = (int) ($_REQUEST['course_id'] ?? 0);
         $exerciseId = $_REQUEST['exercise_id'] ?? 0;
         $status = $_REQUEST['status'] ?? 0;
         if (isset($_GET['filter_by_user']) && !empty($_GET['filter_by_user'])) {
@@ -787,7 +787,7 @@ switch ($action) {
     case 'get_exercise_results':
         $exercise_id = $_REQUEST['exerciseId'];
 
-        $courseId = $_REQUEST['course_id'] ?? 0;
+        $courseId = (int) ($_REQUEST['course_id'] ?? 0);
         $exerciseId = $_REQUEST['exercise_id'] ?? 0;
         $status = $_REQUEST['status'] ?? 0;
         if (isset($_GET['filter_by_user']) && !empty($_GET['filter_by_user'])) {
